Moving tools/cgm2moab over to tools/dagmc.  cgm2moab still built inside dagmc, but this also builds DagMC class and puts that in libdagmc.a (which replaces libcgm2moab.a).
Adding dagmc.make.in/dagmc.make, which adds to the make variables
defined in moab.make to include cgm stuff.  After including moab.make
and dagmc.make, applications only need to use MOAB-defined make
variables to build.  This also ensures applications will get the
version of cgm used to build dagmc under MOAB.

Changed --enable-cgm2moab in configure to --enable-dagmc.

Added generation of a make variable CUBIT_CGM, which adds a 
preprocessor definition of that name if using cgm from cubit.
In this case we have to not call the UW-written RefFace-based ray
fire that isn't part of cubit.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1055 6162379e-bd28-0410-9a7d-b7f4dcfcad3c