Fixing a bug in allocation of sequences.  Before this commit, when: - Existing sequence starts at one place, e.g. 41686, but its sequenceData starts earlier, e.g. 41673, and - You try to allocate a new sequence at a lower handle but with a range which is more than the room available in the sequenceData,
then
- the code wants to allocate right up against the lower end of the
existing sequence, and sizes the new sequenceData accordingly
(i.e. with a max of 41685).

In this case, that new sequenceData will overlap the existing
sequenceData (which starts at 41673, even though it has an empty chunk
in front).

So, this change passes back a sequence data size from the code which
computes where to put the start handle when allocating the new
sequence.

I'm not sure this is the best way to fix this; in the cases I've
observed, this new sequence size is always the number of elements in
the new range.  It may be that you can set that directly when the
returned sequence data pointer is NULL (indicating a new sequence data
must be allocated).  However, this case may only apply when you're
allocating in front of an existing sequence, which probably isn't
always the case.

Also removing unused variable from mbparallelcomm_test, and commented
out ftest stuff in iMesh which was causing the autotesting to fail.

Passed make check.

git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1952 6162379e-bd28-0410-9a7d-b7f4dcfcad3c