Add support for Tripoli4 v7.
Primary changes are to provide more access to metadata and/or support
different metadata.  A method was added to get all the metadata for a
volume, but this probably needs to be replaced by a more generic
lightweight access to tags across the DAGMC interface.

Added support for "composition" definitions that are different from
MCNP materials: cubit group "comp_COMPNAME" (Note: the implicit
complement is assigned a material with "comp_COMPNAME_comp".)

Added support for testing whether a specific surface has a specular or
white reflecting boundary condition and whether a specific volume is
part of the graveyard.

Changed the order of the definition of discard_ and add_ tolerance so
that the add_tolerance can be defined as 100*discard_tolerance as
default.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4017 6162379e-bd28-0410-9a7d-b7f4dcfcad3c