Make HDF5 writer choose dense format for tag data more often.
If at least 2/3 of the entities in an export set have tag values, the tag
is a moab dense tag, and the tag has a default value, then write 
dense-formatted tag data for all entities in the export set.  The reader 
will create a single SequenceData for each export set anyway so no extra 
memory will be allocated to store the tag data when the file is read back 
in.  So this is just a question of using a little more disk space in exchange 
for better IO performance.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4583 6162379e-bd28-0410-9a7d-b7f4dcfcad3c