Fix previous checkin of string arrays for enums: they are not available to C code
If Types.hpp is included in C (not C++) code the declarations appear in the
global namespace rather than the 'moab' one.  However, the actual array
implementations in Types.cpp are always in the 'moab' namespace.  So C code
trying to use them will encounter link errors.  Wrap the array declarations
in Types.hpp in "#ifdef __cpluscplus" so it is obvious at compile time that 
they will not be available to C code.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4703 6162379e-bd28-0410-9a7d-b7f4dcfcad3c