resolves ticket 273 the bug was in intersecting the subranges. For a partial read (that appears in parallel), the id_map between file ids and local entity handles is fragmented, in terms of file ids. When we populate the local sets, we are reading "complete" sets from the file, on each processor, and we have to carefully fish out only the file entities that are found in the id_map (and exist on the local processor) Because of efficiency, we are doing this intersection of subranges using 2 advancing pointers; the construct that crashed was this:  hint = merge.insert( hint, it->value + off, it->value + off + count - 1 );
for some low values of it->value, and negative count, this number  it->value + off + count - 1 was actually negative,
but because it is an EntityHandle, it is an unsigned long int; it became something like 0xfffff...fffd
which is valid :( positive, very large. the set blowed up really bad, because this range that was added was 
ridiculously large. 

two fixes (redundant, in a way): 
1) break out when count becomes negative, and treat the next range in the ranged set. 
2) when advance to a next subrange in id_map, check if we passed the "end", and break out if yes

passes make check in parallel, but I should add a test to capture this bug
it showed up only at reading in parallel, I need to find out how to exercise the partial read of a file in serial 
and manufacture a simpler test


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@6014 6162379e-bd28-0410-9a7d-b7f4dcfcad3c