create polygons, not quads, for MPAS test
interestingly enough, it was working fine with MBQUAD instead of MBPOLYGON
Down deep, in sequence manager, we were creating quads (type 3, MBQUAD), but
with connectivity 5 or 6, so it was a different sequence than the regular
 4-node one for quads.
And in our logic, we were considering the nodes in sequence, as if it were
for a polygon.

So, MOAB can accept 5, or 6-node quads.